% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{calc_vf_internal}
\alias{calc_vf_internal}
\title{Calculate viscous flow at the end of the drainage for plotting}
\usage{
calc_vf_internal(TD, qS, TE, time_vector, exponent = 3/2)
}
\arguments{
\item{TD}{numeric. Arrival time of the drainage front}

\item{qS}{numeric. Either the flux density at the top of the column or the fitted stationary flow.}

\item{TE}{numeric. End of irrigation}

\item{time_vector}{numeric vector. The time vector to calculate the viscous flow.}

\item{exponent}{numeric. The exponent in the equation of viscous flow. Default is 2/3.}
}
\value{
numeric vector. Viscous flow.
}
\description{
This is an internal function. To calculate the viscous flow use calculate_viscous_flow.
}
