% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calculate_drainage_flow}
\alias{calculate_drainage_flow}
\title{Calculates the flow rate from cumulative drainage}
\usage{
calculate_drainage_flow(
  scale_s,
  rho_H20 = 0.99704,
  D,
  delta_out = 30,
  convert_mm_h = TRUE
)
}
\arguments{
\item{scale_s}{tibble. First column time in sec, second column cumulative weight of the drained water}

\item{rho_H20}{numeric. Density of water, default is 0.99704 g/mL or g/cm^3}

\item{D}{numeric. Diameter of the soil column in m.}

\item{delta_out}{numeric. Output sampling time interval, default is 30 sec. Must equal at least the length of the original sampling interval.}

\item{convert_mm_h}{logical. Should the flow rate be converted to mm/h. Default is TRUE.}
}
\value{
tibble. First column is time in sec, second is the flow rate in mm/h
}
\description{
The relative weights are calculated by taking the first derivative using the function D1 from library sfsmisc. Then, they are transformed into a flow rate. Both, a smoothed version of the derivative and the trivial differences are returned.
}
