% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{calculate_qs}
\alias{calculate_qs}
\title{Calculate the flux density at the top of the soil column}
\usage{
calculate_qs(pump_rate, D = 0.08)
}
\arguments{
\item{pump_rate}{numeric. The pump rate in g/min}

\item{D}{numeric. Diameter of the soil column.}
}
\value{
numeric. The flux density \mjeqn{q_S}{} in m/s
}
\description{
Converts the pump rate of the irrigation pump into the flux density \mjeqn{q_S}{} in m/s.
}
