% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{calculate_TW}
\alias{calculate_TW}
\title{Calculate the arrival time of the wetting front}
\usage{
calculate_TW(TD, TE, TB = 0)
}
\arguments{
\item{TD}{numeric. Arrival time of the drainage front.}

\item{TE}{numeric. End of irrigation.}

\item{TB}{numeric. Beginning of irrigation, usually \mjeqn{T_B = 0}{}.}
}
\value{
numeric. Arrival time of the wetting front.
}
\description{
Calculate the arrival time of the wetting front
}
