% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{calculate_vf_parameters}
\alias{calculate_vf_parameters}
\title{Calculate viscous flow parameters}
\usage{
calculate_vf_parameters(eta = 1e-06, g = 9.81, TD, TE, Z, qS)
}
\arguments{
\item{eta}{numeric. Cinetamitc viscosity, default is 1E-6.}

\item{g}{numeric. Gravity constan, default is 9.81 m s\mjeqn{^-2}{}.}

\item{TD}{numeric. Arrival time of the drainage front.}

\item{TE}{numeric. End of irrigation.}

\item{Z}{numeric. Length of the soil column in m.}

\item{qS}{numeric. The flux density at the top of the soil column (or the fitted stationary flow rate) in m s\mjeqn{^-1}{}.}
}
\value{
data.frame
c: celerity
F: film_thickness in \mjeqn{\mu}{}m
L: contact_area in m \mjeqn{^-1}{}
}
\description{
Calculates the parameters of the viscous flow, namely celerity, film thickness and contact area.
}
\examples{
vf_params <- calculate_vf_parameters(TD = 64830, TE = 64410,
Z = 0.3, qS = 2.79E-06)
}
