% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{drainage}
\alias{drainage}
\title{Drainage}
\format{
A data frame with 2184 rows and 2 variables
}
\usage{
data(drainage)
}
\description{
A dataset containing the drainage from a soil column in an irrigation experiment. The variables are the following:
}
\details{
\itemize{
\item time. Time since the start of irrigation in seconds
\item value. The flux in mm/h
}
}
\keyword{datasets}
