% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{extract_data}
\alias{extract_data}
\title{Extracts data from scales}
\usage{
extract_data(data_orig, weight_unit)
}
\arguments{
\item{data_orig}{the original data, the format is "hh:mm:ss.sss:\n weight unit}

\item{weight_unit}{character, weight unit transmitted by the scale: g or kg}
}
\value{
a tibble. First column time in sec, second column weight in unit
}
\description{
Extracts weight data from ascii files generated by HTerm 0.8.1beta
}
