% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{find_tracer_breakthrough}
\alias{find_tracer_breakthrough}
\title{Calculate the breakthrough time of a tracer}
\usage{
find_tracer_breakthrough(
  tracer_data,
  time_interval,
  xlab_tag = "Time (sec)",
  ylab_tag = expression(c/c[0]),
  do_plot = TRUE
)
}
\arguments{
\item{tracer_data}{tibble or data.frame. Tracer breakthrough data.}

\item{time_interval}{vector of length 2. The time interval to search for the breakthrough, typically early in the experiment.}

\item{xlab_tag}{string or expression for the x label. Default is Time (sec).}

\item{ylab_tag}{string or expression for the y label. Default is c/c0.}

\item{do_plot}{logical. Should a plot be produced. Default is TRUE.}
}
\value{
numeric. The estimated time of the tracer breakthrough.
}
\description{
The breakthrough time of a tracer in the drainage of a soil column is calculated as the maximum of curvature of its concentration in a suitable time interval.
}
\examples{
data(tracer)
find_tracer_breakthrough(tracer_data = tracer, time_interval = c(30000, 40000))
}
