% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_drainage_tail}
\alias{fit_drainage_tail}
\title{Fit the viscous flow equation}
\usage{
fit_drainage_tail(
  drainage_data,
  stationary_time,
  D,
  TE,
  TD_interval = NULL,
  qS = NULL,
  fit_qS = TRUE,
  delta_t = 30,
  my_weights = 1
)
}
\arguments{
\item{drainage_data}{tibble or data.frame. Drainage data from a column experiment. The first column must contain the time, the second the drainage data.}

\item{stationary_time}{numeric vector of length 2. The time interval where the flow is assumed to be stationary. Typically one would select the shortly before the irrigation is switched off, i.e. 0.9 * end_of_irrigation until end_or_irrigation.}

\item{D}{numeric. Diameter of the soil column in m.}

\item{TE}{numeric. End of irrigation}

\item{TD_interval}{numeric vector of length 2. The time interval where the arrival time of the drainage front \mjeqn{T_D}{TD} is expected (initial guess for the optimization routine). Default is NULL and is internally replaced by TD_interval = c(0.9 * TE, 1.1 * TE)}

\item{qS}{numeric. Stationary flow rate in m/s. Default is NULL and fit_qS = TRUE to fit \mjeqn{q_S}{qS}, cf. fit_stationary_flow_rate.}

\item{fit_qS}{logical. Should \mjeqn{q_S}{qS} be optimized. Default is TRUE.}

\item{delta_t}{numeric. The time interval in the drainage_data.}

\item{my_weights}{numeric or NULL. Weights for the optimization routine. Experimental, at the moment set to 1.}
}
\value{
list.
q_ms: drainage in m/s
ind: time from which on the fit was done
qS: stationary flow rate in m/s
TD: arrival time of the drainage front
TW: arrival time of the wetting front
TE: end of irrigation (unchanged)
viscous_flow_tail: theoretical drainage according to the viscous flow used for the optimization
}
\description{
Fits the viscous flow equation to the drainage curve of an irrigation experiment on a soil column,
calculates the arrival time of the drainage front \mjeqn{T_D}{TD}.
}
\examples{
data(drainage)
my_TD <- fit_drainage_tail(drainage_data = drainage, stationary_time = c(0.9 * 64410 , 64410),
TE = 64410, TD_interval = c(0.9 * 64410, 1.1 * 64410), qS = NULL, fit_qS = TRUE,
delta_t = 30, my_weights = NULL)
}
