% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_stationary_flow_rate}
\alias{fit_stationary_flow_rate}
\title{Calculate stationary flow rate in drainage}
\usage{
fit_stationary_flow_rate(drainage_data, time_interval)
}
\arguments{
\item{drainage_data}{tibble or data.frame. Drainage data from a column experiment. First column must be time, second the drainage data.}

\item{time_interval}{numeric vector of length 2. The time interval where the flow is assumed to be stationary. Typically one would select the shortly before the irrigation is switched off, i.e. 0.9 * end_of_irrigation until end_or_irrigation.}
}
\value{
numeric. The stationary flow rate
}
\description{
Calculate stationary flow rate in drainage
}
\examples{
data(drainage)
fit_stationary_flow_rate(drainage_data = drainage, time_interval = c(0.9 * 64404 , 64404))
}
