% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_tracer}
\alias{plot_tracer}
\title{Plot the tracer breakthrough curve}
\usage{
plot_tracer(
  tracer_data,
  time_threshold = NULL,
  file_name = NULL,
  save_pdf = FALSE
)
}
\arguments{
\item{tracer_data}{tibble or data.frame. Tracer breakthrough data.
The fist column should contain the time and the second the tracer concentration.}

\item{time_threshold}{numeric. Time of the breakthrough (i.e. start of the concentration increase)}

\item{file_name}{string. The name for the pdf file to save the figure. Default is NULL}

\item{save_pdf}{logical. Default is FALSE}
}
\value{
ggplot object
}
\description{
Plot the tracer breakthrough curve
}
\examples{
plot_tracer(tracer_data = tracer, time_threshold = 40000)
}
