% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{tracer}
\alias{tracer}
\title{Tracer}
\format{
A data frame with 213 rows and 2 variables
}
\usage{
data(tracer)
}
\description{
A dataset containing the breakthrough curve of bromide. The variables are the following:
}
\details{
\itemize{
\item time. Time since the start of irrigation in seconds
\item value. The relative tracer concentration c/c$_0$ (divided by the concentration in the irrigation water)
}
}
\keyword{datasets}
