% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_stationary_flow_rate}
\alias{fit_stationary_flow_rate}
\title{Calculate stationary flow rate in the drainage}
\usage{
fit_stationary_flow_rate(drainage_data, stationary_time)
}
\arguments{
\item{drainage_data}{tibble or data.frame. Drainage data from a column experiment. The first column must contain the time, the second the drainage data.}

\item{stationary_time}{numeric vector of length 2. The time interval where the flow is assumed to be stationary. Typically one would select the shortly before the irrigation is switched off, i.e. 0.9 * end_of_irrigation until end_or_irrigation.}
}
\value{
numeric. The stationary flow rate
}
\description{
Calculates the median of the flow rate in a given time interval, usually the stationary flow interval.
}
\examples{
data(drainage)
fit_stationary_flow_rate(drainage_data = drainage, stationary_time = c(0.9 * 64410 , 64410))
}
