% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_tracer}
\alias{plot_tracer}
\title{Plot the tracer breakthrough curve}
\usage{
plot_tracer(
  tracer_data,
  time_threshold = NULL,
  xlab_tag = "Time (sec)",
  ylab_tag = expression(c/c[0]),
  file_name = NULL,
  save_pdf = FALSE
)
}
\arguments{
\item{tracer_data}{tibble or data.frame. Tracer breakthrough data.
The fist column should contain the time and the second the tracer concentration.}

\item{time_threshold}{numeric. Time of the breakthrough (i.e. start of the concentration increase)}

\item{xlab_tag}{string or expression for the x label. Default is Time (sec).}

\item{ylab_tag}{string or expression for the y label. Default is c/c0.}

\item{file_name}{string. The name for the pdf file to save the figure. Default is NULL.}

\item{save_pdf}{logical. Should de figure be saved to a pdf file. Default is FALSE.}
}
\value{
ggplot object
}
\description{
Plot the tracer breakthrough curve
}
\examples{
data(tracer)
plot_tracer(tracer_data = tracer, time_threshold = 40000)
}
